Renamed Extension Methods for enums. Added functionality to download files asynchronously from Hattrick. Added missing documentation. Added functionality to process TeamDetails Xml file. Extended WorldDetails Xml file process functionality. Added LeagueCupTier and LeagueCupType enums. Added new database business objects, mappings and related changes. Added TeamDetails Xml file business objects and related changes. Modified WorldDetails Cup and League business objects. Added new constants. Added new Xml file parser and improved existing ones. Added new localized strings. Added new FormGenericProgress to show asynchronous tasks progress. Added new message box for allowed token. Added download tasks and control population functionality to FormUser

+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/WorldDetails/Cup.cs
@@ -25,7 +25,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.WorldDetails
         /// For National cups (Levels 1 through 6) DO NOT have a League Level, and Hattrick
         /// represents this with a 0 instead of a NULL.
         /// </remarks>
-        public int CupLeagueLevel { get; set; }
+        public byte CupLeagueLevel { get; set; }
 
         /// <summary>
         /// Gets or sets the Cup Level.
@@ -34,7 +34,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.WorldDetails
         /// DO NOT mistake with CupLeagueLevel. National/Divisional Cups: 1. Challenger Cups: 2.
         /// Consolation Cups: 3.
         /// </remarks>
-        public int CupLevel { get; set; }
+        public byte CupLevel { get; set; }
 
         /// <summary>
         /// Gets or sets the CupLevelIndex
@@ -45,7 +45,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.WorldDetails
         /// Ruby: 2.
         /// Sapphire: 3.
         /// </remarks>
-        public int CupLevelIndex { get; set; }
+        public byte CupLevelIndex { get; set; }
 
         /// <summary>
         /// Gets or sets the name of the Cup.
@@ -55,12 +55,12 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.WorldDetails
         /// <summary>
         /// Gets or sets the current match round of the Cup.
         /// </summary>
-        public int MatchRound { get; set; }
+        public byte MatchRound { get; set; }
 
         /// <summary>
         /// Gets or sets the number of match rounds of the Cup left to be played.
         /// </summary>
-        public int MatchRoundsLeft { get; set; }
+        public byte MatchRoundsLeft { get; set; }
 
         #endregion Public Properties
     }
